+++ packages/web3x/example/lib/erc20.g.dart
@@ -52,6 +52,7 @@ class Erc20 extends _i1.GeneratedContract {
     BigInt amount, {
     required _i1.Credentials credentials,
     _i1.Transaction? transaction,
+    _i1.Transaction? additional,
   }) async {
     final function = self.abi.functions[1];
     assert(checkSignature(function, '095ea7b3'));
@@ -60,10 +61,11 @@ class Erc20 extends _i1.GeneratedContract {
       amount,
     ];
     return write(
-      credentials,
-      transaction,
-      function,
-      params,
+      credentials: credentials,
+      base: transaction,
+      additional: additional,
+      function: function,
+      parameters: params,
     );
   }
 
@@ -153,6 +155,7 @@ class Erc20 extends _i1.GeneratedContract {
     BigInt amount, {
     required _i1.Credentials credentials,
     _i1.Transaction? transaction,
+    _i1.Transaction? additional,
   }) async {
     final function = self.abi.functions[7];
     assert(checkSignature(function, 'a9059cbb'));
@@ -161,10 +164,11 @@ class Erc20 extends _i1.GeneratedContract {
       amount,
     ];
     return write(
-      credentials,
-      transaction,
-      function,
-      params,
+      credentials: credentials,
+      base: transaction,
+      additional: additional,
+      function: function,
+      parameters: params,
     );
   }
 
@@ -177,6 +181,7 @@ class Erc20 extends _i1.GeneratedContract {
     BigInt amount, {
     required _i1.Credentials credentials,
     _i1.Transaction? transaction,
+    _i1.Transaction? additional,
   }) async {
     final function = self.abi.functions[8];
     assert(checkSignature(function, '23b872dd'));
@@ -186,10 +191,11 @@ class Erc20 extends _i1.GeneratedContract {
       amount,
     ];
     return write(
-      credentials,
-      transaction,
-      function,
-      params,
+      credentials: credentials,
+      base: transaction,
+      additional: additional,
+      function: function,
+      parameters: params,
     );
   }
 
@@ -210,7 +216,10 @@ class Erc20 extends _i1.GeneratedContract {
         result.topics!,
         result.data!,
       );
-      return Approval(decoded);
+      return Approval(
+        decoded,
+        result,
+      );
     });
   }
 
@@ -231,14 +240,19 @@ class Erc20 extends _i1.GeneratedContract {
         result.topics!,
         result.data!,
       );
-      return Transfer(decoded);
+      return Transfer(
+        decoded,
+        result,
+      );
     });
   }
 }
 
 class Approval {
-  Approval(List<dynamic> response)
-      : owner = (response[0] as _i1.EthereumAddress),
+  Approval(
+    List<dynamic> response,
+    this.event,
+  )   : owner = (response[0] as _i1.EthereumAddress),
         spender = (response[1] as _i1.EthereumAddress),
         value = (response[2] as BigInt);
 
@@ -247,11 +261,15 @@ class Approval {
   final _i1.EthereumAddress spender;
 
   final BigInt value;
+
+  final _i1.FilterEvent event;
 }
 
 class Transfer {
-  Transfer(List<dynamic> response)
-      : from = (response[0] as _i1.EthereumAddress),
+  Transfer(
+    List<dynamic> response,
+    this.event,
+  )   : from = (response[0] as _i1.EthereumAddress),
         to = (response[1] as _i1.EthereumAddress),
         value = (response[2] as BigInt);
 
@@ -260,4 +278,6 @@ class Transfer {
   final _i1.EthereumAddress to;
 
   final BigInt value;
+
+  final _i1.FilterEvent event;
 }