+++ packages/web3x/lib/src/utils/extensions.dart
@@ -1,0 +1,8 @@
+extension StringExtension on String {
+  String stripOx() {
+    if (startsWith('0x')) {
+      return substring(2);
+    }
+    return this;
+  }
+}