+++ packages/web3x_builders/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Add an `additional` parameter for calling contract modification methods.
+
 ## 1.0.0
 
 - Clone from [https://github.com/inveker/web3dart_builders](https://github.com/inveker/web3dart_builders)